Load last entry on boot
Some items are set before initialization, so we
need to manually add these instead of only relying
on the changelisteners.

Strangely enough I ran into issues where the
application would immediately shut down without
throwing any exceptions if I tried to set the
appdir directory in the AppdirController after
starting the stage.
So I've left it as-is, we set it before starting
the stage and update manually.
Might be worthwhile to investigate why this is
however, ideally we'd have all controllers
initialized first before loading anything from
file.