Allow (de)selecting Authors and Collections
The changes are applied to an underlying set
representing the filtered/selected items which
have both a selected collection and at least one
selected author.

However, the listview in the view does not seem
to reflect the changes in the list.
We have double checked this with a change
listener for the size, so we know for sure that
the underlying list changes.